从gorilla/rpc/v2中独立，新增RegisterMethod, RegisterAlias
保留版权声明，协议与gorilla/rpc一致

- 增加RegisterMethod，支持完整路径定义method
- 增加RegisterAlias，支持多路径映射同一方法
- 按jsonrpc规范，修订Error/Reply，当成功时error不输出，当失败时reply不输出
- 修订测试用例
- 将testapp从json中剥离，形成独立的测试应用目录
- 将WriteError从server.go文件中独立成writer.go
- 将codec接口定义从server.go文件中独立成codec.go
