+++ README.md
@@ -1,7 +1,9 @@
 rpc
 ===
-[![Build Status](https://travis-ci.org/gorilla/rpc.png?branch=master)](https://travis-ci.org/gorilla/rpc)
+
+
+rpc/v2 support for JSON-RPC 2.0 Specification.
 
 gorilla/rpc is a foundation for RPC over HTTP services, providing access to the exported methods of an object through HTTP requests.
 
-Read the full documentation here: https://www.gorillatoolkit.org/pkg/rpc
+Read the full documentation here: http://www.gorillatoolkit.org/pkg/rpc
