+++ compression_selector.go
