+++ go.mod
@@ -1,0 +1,3 @@
+module wbylovesun/rpc
+
+go 1.18
