+++ json/client.go
@@ -7,8 +7,6 @@ package json
 
 import (
 	"encoding/json"
-	"errors"
-	"fmt"
 	"io"
 	"math/rand"
 )
@@ -19,10 +17,15 @@ import (
 
 // clientRequest represents a JSON-RPC request sent by a client.
 type clientRequest struct {
+	// JSON-RPC protocol.
+	Version string `json:"jsonrpc"`
+
 	// A String containing the name of the method to be invoked.
 	Method string `json:"method"`
+
 	// Object to pass as request parameter to the method.
-	Params [1]interface{} `json:"params"`
+	Params interface{} `json:"params"`
+
 	// The request id. This can be of any type. It is used to match the
 	// response with the request that it is replying to.
 	Id uint64 `json:"id"`
@@ -30,17 +33,18 @@ type clientRequest struct {
 
 // clientResponse represents a JSON-RPC response returned to a client.
 type clientResponse struct {
-	Result *json.RawMessage `json:"result"`
-	Error  interface{}      `json:"error"`
-	Id     uint64           `json:"id"`
+	Version string           `json:"jsonrpc"`
+	Result  *json.RawMessage `json:"result"`
+	Error   *json.RawMessage `json:"error"`
 }
 
 // EncodeClientRequest encodes parameters for a JSON-RPC client request.
 func EncodeClientRequest(method string, args interface{}) ([]byte, error) {
 	c := &clientRequest{
-		Method: method,
-		Params: [1]interface{}{args},
-		Id:     uint64(rand.Int63()),
+		Version: "2.0",
+		Method:  method,
+		Params:  args,
+		Id:      uint64(rand.Int63()),
 	}
 	return json.Marshal(c)
 }
@@ -53,10 +57,19 @@ func DecodeClientResponse(r io.Reader, reply interface{}) error {
 		return err
 	}
 	if c.Error != nil {
-		return fmt.Errorf("%v", c.Error)
+		jsonErr := &Error{}
+		if err := json.Unmarshal(*c.Error, jsonErr); err != nil {
+			return &Error{
+				Code:    E_SERVER,
+				Message: string(*c.Error),
+			}
+		}
+		return jsonErr
 	}
+
 	if c.Result == nil {
-		return errors.New("result is null")
+		return ErrNullResult
 	}
+
 	return json.Unmarshal(*c.Result, reply)
 }
