+++ map.go
@@ -26,14 +26,13 @@ var (
 // ----------------------------------------------------------------------------
 
 type service struct {
-	name     string                    // name of service
-	rcvr     reflect.Value             // receiver of methods for the service
-	rcvrType reflect.Type              // type of the receiver
-	methods  map[string]*serviceMethod // registered methods
-	passReq  bool
+	name     string        // name of service
+	rcvr     reflect.Value // receiver of methods for the service
+	rcvrType reflect.Type  // type of the receiver
 }
 
 type serviceMethod struct {
+	service   *service       // service
 	method    reflect.Method // receiver method
 	argsType  reflect.Type   // type of the request argument
 	replyType reflect.Type   // type of the response argument
@@ -45,19 +44,19 @@ type serviceMethod struct {
 
 // serviceMap is a registry for services.
 type serviceMap struct {
-	mutex    sync.Mutex
-	services map[string]*service
+	mutex   sync.Mutex
+	methods map[string]*serviceMethod
+	aliases map[string]string
 }
 
 // register adds a new service using reflection to extract its methods.
-func (m *serviceMap) register(rcvr interface{}, name string, passReq bool) error {
+func (m *serviceMap) register(rcvr interface{}, name string) error {
 	// Setup service.
 	s := &service{
 		name:     name,
 		rcvr:     reflect.ValueOf(rcvr),
 		rcvrType: reflect.TypeOf(rcvr),
-		methods:  make(map[string]*serviceMethod),
-		passReq:  passReq,
+		//methods:  make(map[string]*serviceMethod),
 	}
 	if name == "" {
 		s.name = reflect.Indirect(s.rcvr).Type().Name()
@@ -70,43 +69,30 @@ func (m *serviceMap) register(rcvr interface{}, name string, passReq bool) error
 			s.rcvrType.String())
 	}
 	// Setup methods.
+	methods := map[string]*serviceMethod{}
 	for i := 0; i < s.rcvrType.NumMethod(); i++ {
 		method := s.rcvrType.Method(i)
 		mtype := method.Type
-
-		// offset the parameter indexes by one if the
-		// service methods accept an HTTP request pointer
-		var paramOffset int
-		if passReq {
-			paramOffset = 1
-		} else {
-			paramOffset = 0
-		}
-
 		// Method must be exported.
 		if method.PkgPath != "" {
 			continue
 		}
 		// Method needs four ins: receiver, *http.Request, *args, *reply.
-		if mtype.NumIn() != 3+paramOffset {
+		if mtype.NumIn() != 4 {
 			continue
 		}
-
-		// If the service methods accept an HTTP request pointer
-		if passReq {
-			// First argument must be a pointer and must be http.Request.
-			reqType := mtype.In(1)
-			if reqType.Kind() != reflect.Ptr || reqType.Elem() != typeOfRequest {
-				continue
-			}
+		// First argument must be a pointer and must be http.Request.
+		reqType := mtype.In(1)
+		if reqType.Kind() != reflect.Ptr || reqType.Elem() != typeOfRequest {
+			continue
 		}
-		// Next argument must be a pointer and must be exported.
-		args := mtype.In(1 + paramOffset)
+		// Second argument must be a pointer and must be exported.
+		args := mtype.In(2)
 		if args.Kind() != reflect.Ptr || !isExportedOrBuiltin(args) {
 			continue
 		}
-		// Next argument must be a pointer and must be exported.
-		reply := mtype.In(2 + paramOffset)
+		// Third argument must be a pointer and must be exported.
+		reply := mtype.In(3)
 		if reply.Kind() != reflect.Ptr || !isExportedOrBuiltin(reply) {
 			continue
 		}
@@ -117,25 +103,107 @@ func (m *serviceMap) register(rcvr interface{}, name string, passReq bool) error
 		if returnType := mtype.Out(0); returnType != typeOfError {
 			continue
 		}
-		s.methods[method.Name] = &serviceMethod{
+		methodName, err := toLowerCase(method.Name)
+		if err != nil {
+			continue
+		}
+		methodName = s.name + "/" + methodName
+		methods[methodName] = &serviceMethod{
+			service:   s,
 			method:    method,
 			argsType:  args.Elem(),
 			replyType: reply.Elem(),
 		}
 	}
-	if len(s.methods) == 0 {
-		return fmt.Errorf("rpc: %q has no exported methods of suitable type",
-			s.name)
+	if len(methods) == 0 {
+		return fmt.Errorf(
+			"rpc: %q has no exported methods of suitable type",
+			s.name,
+		)
 	}
 	// Add to the map.
 	m.mutex.Lock()
 	defer m.mutex.Unlock()
-	if m.services == nil {
-		m.services = make(map[string]*service)
-	} else if _, ok := m.services[s.name]; ok {
-		return fmt.Errorf("rpc: service already defined: %q", s.name)
+	for name, method := range methods {
+		if _, ok := m.methods[name]; ok {
+			return fmt.Errorf("rpc: service method already defined: %q", name)
+		}
+		m.methods[name] = method
 	}
-	m.services[s.name] = s
+	return nil
+}
+
+// register method for specific service that does not export all methods with method full name.
+func (m *serviceMap) registerMethod(rcvr any, name string, methodName string) error {
+	if name == "" {
+		return fmt.Errorf("rpc: service method name must not be empty")
+	}
+	// Setup service.
+	s := &service{
+		rcvr:     reflect.ValueOf(rcvr),
+		rcvrType: reflect.TypeOf(rcvr),
+	}
+	s.name = reflect.Indirect(s.rcvr).Type().Name()
+	method, exists := s.rcvrType.MethodByName(methodName)
+	if !exists {
+		return fmt.Errorf("rpc: service method not found: %q", methodName)
+	}
+	mtype := method.Type
+	// Method must be exported.
+	if method.PkgPath != "" {
+		return fmt.Errorf("rpc: service method PkgPath %s is not empty", method.PkgPath)
+	}
+	// Method needs four ins: receiver, *http.Request, *args, *reply.
+	if mtype.NumIn() != 4 {
+		return fmt.Errorf("rpc: service method NumIn %d is not 4", mtype.NumIn())
+	}
+	// First argument must be a pointer and must be http.Request.
+	reqType := mtype.In(1)
+	if reqType.Kind() != reflect.Ptr || reqType.Elem() != typeOfRequest {
+		return fmt.Errorf("rpc: service method first arg must be an http Request Pointer")
+	}
+	// Second argument must be a pointer and must be exported.
+	args := mtype.In(2)
+	if args.Kind() != reflect.Ptr || !isExportedOrBuiltin(args) {
+		return fmt.Errorf("rpc: service method second arg must be an Pointer")
+	}
+	// Third argument must be a pointer and must be exported.
+	reply := mtype.In(3)
+	if reply.Kind() != reflect.Ptr || !isExportedOrBuiltin(reply) {
+		return fmt.Errorf("rpc: service method third arg must be an Pointer")
+	}
+	// Method needs one out: error.
+	if mtype.NumOut() != 1 {
+		return fmt.Errorf("rpc: service method NumOut must be 1, now is %d", mtype.NumOut())
+	}
+	if returnType := mtype.Out(0); returnType != typeOfError {
+		return fmt.Errorf("rpc: service method typeof NumOut must be Error")
+	}
+	m.mutex.Lock()
+	defer m.mutex.Unlock()
+	if _, ok := m.methods[name]; ok {
+		return fmt.Errorf("rpc: service method already defined: %q", name)
+	}
+	m.methods[name] = &serviceMethod{
+		service:   s,
+		method:    method,
+		argsType:  args.Elem(),
+		replyType: reply.Elem(),
+	}
+	return nil
+}
+
+// allow to use different names to call same rpc method.
+func (m *serviceMap) registerAlias(alias, target string) error {
+	if _, ok := m.methods[target]; !ok {
+		return fmt.Errorf("rpc: service method %s for alias not found", target)
+	}
+	if _, ok := m.aliases[alias]; ok {
+		return fmt.Errorf("rpc: service method alias %s already defined", alias)
+	}
+	m.mutex.Lock()
+	defer m.mutex.Unlock()
+	m.aliases[alias] = target
 	return nil
 }
 
@@ -143,21 +211,25 @@ func (m *serviceMap) register(rcvr interface{}, name string, passReq bool) error
 //
 // The method name uses a dotted notation as in "Service.Method".
 func (m *serviceMap) get(method string) (*service, *serviceMethod, error) {
-	parts := strings.Split(method, ".")
-	if len(parts) != 2 {
-		err := fmt.Errorf("rpc: service/method request ill-formed: %q", method)
+	//parts := strings.Split(method, ".")
+	//if len(parts) != 2 {
+	//	err := fmt.Errorf("rpc: service/method request ill-formed: %q", method)
+	//	return nil, nil, err
+	//}
+	m.mutex.Lock()
+	target, ok := m.aliases[method]
+	if ok {
+		method = target
+	}
+	serviceMethod := m.methods[method]
+	m.mutex.Unlock()
+	if serviceMethod == nil {
+		err := fmt.Errorf("rpc: can't find service method %q", method)
 		return nil, nil, err
 	}
-	m.mutex.Lock()
-	service := m.services[parts[0]]
-	m.mutex.Unlock()
+	service := serviceMethod.service
 	if service == nil {
 		err := fmt.Errorf("rpc: can't find service %q", method)
-		return nil, nil, err
-	}
-	serviceMethod := service.methods[parts[1]]
-	if serviceMethod == nil {
-		err := fmt.Errorf("rpc: can't find method %q", method)
 		return nil, nil, err
 	}
 	return service, serviceMethod, nil
@@ -177,4 +249,30 @@ func isExportedOrBuiltin(t reflect.Type) bool {
 	// PkgPath will be non-empty even for an exported type,
 	// so we need to check the type name as well.
 	return isExported(t.Name()) || t.PkgPath() == ""
+}
+
+func isValidRouteName(s string) bool {
+	isASCII, hasUpper := true, false
+	for i := 0; i < len(s); i++ {
+		c := s[i]
+		if c >= utf8.RuneSelf {
+			isASCII = false
+			break
+		}
+		hasUpper = hasUpper || ('A' <= c && c <= 'Z')
+	}
+	return isASCII
+}
+
+func toLowerCase(s string) (string, error) {
+	if !isValidRouteName(s) {
+		return "", fmt.Errorf("invalid route name %s", s)
+	}
+	if len(s) == 0 {
+		return "", nil
+	}
+	if len(s) == 1 {
+		return strings.ToLower(s), nil
+	}
+	return strings.ToLower(string(s[0])) + s[1:], nil
 }
