+++ server_test.go
@@ -6,34 +6,38 @@
 package rpc
 
 import (
+	"errors"
 	"net/http"
 	"strconv"
 	"testing"
 )
 
-type Service1Request struct {
+type ServiceRequest struct {
 	A int
 	B int
 }
 
-type Service1Response struct {
+type ServiceResponse struct {
 	Result int
 }
 
 type Service1 struct {
 }
 
-func (t *Service1) Multiply(r *http.Request, req *Service1Request, res *Service1Response) error {
+func (t *Service1) Multiply(r *http.Request, req *ServiceRequest, res *ServiceResponse) error {
 	res.Result = req.A * req.B
 	return nil
 }
 
-func (t *Service1) Add(req *Service1Request, res *Service1Response) error {
-	res.Result = req.A + req.B
-	return nil
+type Service2 struct {
 }
 
-type Service2 struct {
+type Service3 struct {
+}
+
+func (t *Service3) Add(r *http.Request, req *ServiceRequest, res *ServiceResponse) error {
+	res.Result = req.A + req.B
+	return nil
 }
 
 func TestRegisterService(t *testing.T) {
@@ -44,12 +48,12 @@ func TestRegisterService(t *testing.T) {
 
 	// Inferred name.
 	err = s.RegisterService(service1, "")
-	if err != nil || !s.HasMethod("Service1.Multiply") {
+	if err != nil || !s.HasMethod("Service1/multiply") {
 		t.Errorf("Expected to be registered: Service1.Multiply")
 	}
 	// Provided name.
 	err = s.RegisterService(service1, "Foo")
-	if err != nil || !s.HasMethod("Foo.Multiply") {
+	if err != nil || !s.HasMethod("Foo/multiply") {
 		t.Errorf("Expected to be registered: Foo.Multiply")
 	}
 	// No methods.
@@ -59,60 +63,39 @@ func TestRegisterService(t *testing.T) {
 	}
 }
 
-func TestRegisterTCPService(t *testing.T) {
-	var err error
-	s := NewServer()
-	service1 := new(Service1)
-	service2 := new(Service2)
-
-	// Inferred name.
-	err = s.RegisterTCPService(service1, "")
-	if err != nil || !s.HasMethod("Service1.Add") {
-		t.Errorf("Expected to be registered: Service1.Add")
-	}
-	// Provided name.
-	err = s.RegisterTCPService(service1, "Foo")
-	if err != nil || !s.HasMethod("Foo.Add") {
-		t.Errorf("Expected to be registered: Foo.Add")
-	}
-	// No methods.
-	err = s.RegisterTCPService(service2, "")
-	if err == nil {
-		t.Errorf("Expected error on service2")
-	}
-}
-
 // MockCodec decodes to Service1.Multiply.
 type MockCodec struct {
-	A, B int
+	A, B   int
+	Method string
 }
 
-func (c MockCodec) NewRequest(*http.Request) CodecRequest {
-	return MockCodecRequest{c.A, c.B}
+func (c MockCodec) NewRequest(*http.Request) Request {
+	return MockCodecRequest{c.A, c.B, c.Method}
 }
 
 type MockCodecRequest struct {
-	A, B int
+	A, B       int
+	MethodName string
 }
 
 func (r MockCodecRequest) Method() (string, error) {
-	return "Service1.Multiply", nil
+	return r.MethodName, nil
 }
 
 func (r MockCodecRequest) ReadRequest(args interface{}) error {
-	req := args.(*Service1Request)
+	req := args.(*ServiceRequest)
 	req.A, req.B = r.A, r.B
 	return nil
 }
 
-func (r MockCodecRequest) WriteResponse(w http.ResponseWriter, reply interface{}, methodErr error) error {
-	if methodErr != nil {
-		w.Write([]byte(methodErr.Error()))
-	} else {
-		res := reply.(*Service1Response)
-		w.Write([]byte(strconv.Itoa(res.Result)))
-	}
-	return nil
+func (r MockCodecRequest) WriteResponse(w http.ResponseWriter, reply interface{}) {
+	res := reply.(*ServiceResponse)
+	w.Write([]byte(strconv.Itoa(res.Result)))
+}
+
+func (r MockCodecRequest) WriteError(w http.ResponseWriter, status int, err error) {
+	w.WriteHeader(status)
+	w.Write([]byte(err.Error()))
 }
 
 type MockResponseWriter struct {
@@ -151,8 +134,7 @@ func TestServeHTTP(t *testing.T) {
 
 	s := NewServer()
 	s.RegisterService(new(Service1), "")
-	s.RegisterCodec(MockCodec{A, B}, "mock")
-
+	s.RegisterCodec(MockCodec{A, B, "Service1/multiply"}, "mock")
 	r, err := http.NewRequest("POST", "", nil)
 	if err != nil {
 		t.Fatal(err)
@@ -204,15 +186,137 @@ func TestInterception(t *testing.T) {
 
 	s := NewServer()
 	s.RegisterService(new(Service1), "")
-	s.RegisterCodec(MockCodec{A, B}, "mock")
+	s.RegisterCodec(MockCodec{A, B, "Service1/multiply"}, "mock")
 	s.RegisterInterceptFunc(func(i *RequestInfo) *http.Request {
 		return r2
 	})
+	s.RegisterValidateRequestFunc(func(info *RequestInfo, v interface{}) error { return nil })
 	s.RegisterAfterFunc(func(i *RequestInfo) {
 		if i.Request != r2 {
 			t.Errorf("Request was %v, should be %v.", i.Request, r2)
 		}
 	})
+
+	r, err := http.NewRequest("POST", "", nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+	r.Header.Set("Content-Type", "mock; dummy")
+	w := NewMockResponseWriter()
+	s.ServeHTTP(w, r)
+	if w.Status != 200 {
+		t.Errorf("Status was %d, should be 200.", w.Status)
+	}
+	if w.Body != strconv.Itoa(expected) {
+		t.Errorf("Response body was %s, should be %s.", w.Body, strconv.Itoa(expected))
+	}
+}
+func TestValidationSuccessful(t *testing.T) {
+	const (
+		A = 2
+		B = 3
+
+		expected = A * B
+	)
+
+	validate := func(info *RequestInfo, v interface{}) error { return nil }
+
+	s := NewServer()
+	s.RegisterService(new(Service1), "")
+	s.RegisterCodec(MockCodec{A, B, "Service1/multiply"}, "mock")
+	s.RegisterValidateRequestFunc(validate)
+
+	r, err := http.NewRequest("POST", "", nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+	r.Header.Set("Content-Type", "mock; dummy")
+	w := NewMockResponseWriter()
+	s.ServeHTTP(w, r)
+	if w.Status != 200 {
+		t.Errorf("Status was %d, should be 200.", w.Status)
+	}
+	if w.Body != strconv.Itoa(expected) {
+		t.Errorf("Response body was %s, should be %s.", w.Body, strconv.Itoa(expected))
+	}
+}
+
+func TestValidationFails(t *testing.T) {
+	const expected = "this instance only supports zero values"
+
+	validate := func(r *RequestInfo, v interface{}) error {
+		req := v.(*ServiceRequest)
+		if req.A != 0 || req.B != 0 {
+			return errors.New(expected)
+		}
+		return nil
+	}
+
+	s := NewServer()
+	s.RegisterService(new(Service1), "")
+	s.RegisterCodec(MockCodec{1, 2, "Service1/multiply"}, "mock")
+	s.RegisterValidateRequestFunc(validate)
+
+	r, err := http.NewRequest("POST", "", nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+	r.Header.Set("Content-Type", "mock; dummy")
+	w := NewMockResponseWriter()
+	s.ServeHTTP(w, r)
+	if w.Status != 400 {
+		t.Errorf("Status was %d, should be 200.", w.Status)
+	}
+	if w.Body != expected {
+		t.Errorf("Response body was %s, should be %s.", w.Body, expected)
+	}
+}
+
+func TestRegisterMethodSuccessful(t *testing.T) {
+	const (
+		A = 2
+		B = 3
+
+		expected = A + B
+	)
+
+	validate := func(info *RequestInfo, v interface{}) error { return nil }
+
+	s := NewServer()
+	s.RegisterMethod(new(Service3), "v1/calc/sum2", "Add")
+	s.RegisterCodec(MockCodec{A, B, "v1/calc/sum2"}, "mock")
+	s.RegisterValidateRequestFunc(validate)
+
+	r, err := http.NewRequest("POST", "", nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+	r.Header.Set("Content-Type", "mock; dummy")
+	w := NewMockResponseWriter()
+	s.ServeHTTP(w, r)
+	if w.Status != 200 {
+		t.Errorf("Status was %d, should be 200.", w.Status)
+	}
+	if w.Body != strconv.Itoa(expected) {
+		t.Errorf("Response body was %s, should be %s.", w.Body, strconv.Itoa(expected))
+	}
+}
+
+func TestAliasSuccessful(t *testing.T) {
+	const (
+		A = 2
+		B = 3
+
+		expected = A + B
+	)
+
+	validate := func(info *RequestInfo, v interface{}) error { return nil }
+
+	s := NewServer()
+	s.RegisterMethod(new(Service3), "v1/calc/sum2", "Add")
+	s.RegisterAlias("v2/calc/sum", "v1/calc/sum2")
+	s.RegisterCodec(MockCodec{A, B, "v2/calc/sum"}, "mock")
+	s.RegisterValidateRequestFunc(validate)
 
 	r, err := http.NewRequest("POST", "", nil)
 	if err != nil {
