+++ testapp/counter.go
@@ -9,9 +9,8 @@ import (
 	"flag"
 	"log"
 	"net/http"
-
-	"github.com/gorilla/rpc/v2"
-	"github.com/gorilla/rpc/v2/json2"
+	"wbylovesun/rpc"
+	"wbylovesun/rpc/json"
 )
 
 type Counter struct {
@@ -22,8 +21,8 @@ type IncrReq struct {
 	Delta int
 }
 
-// Notification.
-func (c *Counter) Incr(r *http.Request, req *IncrReq, res *json2.EmptyResponse) error {
+// Incr Notification.
+func (c *Counter) Incr(r *http.Request, req *IncrReq, res *json.EmptyResponse) error {
 	log.Printf("<- Incr %+v", *req)
 	c.Count += req.Delta
 	return nil
@@ -40,9 +39,9 @@ func (c *Counter) Get(r *http.Request, req *GetReq, res *Counter) error {
 }
 
 func main() {
-	address := flag.String("address", ":65534", "")
+	address := flag.String("address", ":63342", "")
 	s := rpc.NewServer()
-	s.RegisterCodec(json2.NewCustomCodec(&rpc.CompressionSelector{}), "application/json")
+	s.RegisterCodec(json.NewCustomCodec(&rpc.CompressionSelector{}), "application/json")
 	s.RegisterService(new(Counter), "")
 	http.Handle("/", http.StripPrefix("/", http.FileServer(http.Dir("./"))))
 	http.Handle("/jsonrpc/", s)
