--- v2/README.md
@@ -1,9 +1,0 @@
-rpc
-===
-
-
-rpc/v2 support for JSON-RPC 2.0 Specification.
-
-gorilla/rpc is a foundation for RPC over HTTP services, providing access to the exported methods of an object through HTTP requests.
-
-Read the full documentation here: http://www.gorillatoolkit.org/pkg/rpc
